Implement the basic library
That connects Tactician to Zend Expressive (using SF dependency
injection compiler passes).

A route will be created according to the "http_route" type defined on
the DI container tag used to define the command/query handler.

The handler will be automatically mapped to the right message bus
defined on the "bus.handler" tag.